feat: エントリーポイントとAppクラスのファイルを分離
